Rename to rollup transform generator
