+++ x-pack/plugins/observability/server/services/slo/rollup_transform_generators/common.ts
