+++ x-pack/plugins/observability/server/services/slo/transform_manager.ts
@@ -9,7 +9,7 @@ import { ElasticsearchClient, Logger } from '@kbn/core/server';
 
 import { SLO, IndicatorTypes } from '../../domain/models';
 import { retryTransientEsErrors } from '../../utils/retry';
-import { TransformGenerator } from './transform_generators';
+import { RollupTransformGenerator } from './rollup_transform_generators';
 
 type TransformId = string;
 
@@ -22,25 +22,25 @@ export interface TransformManager {
 
 export class DefaultTransformManager implements TransformManager {
   constructor(
-    private generators: Record<IndicatorTypes, TransformGenerator>,
+    private rollupGenerators: Record<IndicatorTypes, RollupTransformGenerator>,
     private esClient: ElasticsearchClient,
     private logger: Logger
   ) {}
 
   async install(slo: SLO): Promise<TransformId> {
-    const generator = this.generators[slo.indicator.type];
-    if (!generator) {
+    const rollupGenerator = this.rollupGenerators[slo.indicator.type];
+    if (!rollupGenerator) {
       this.logger.error(`No transform generator found for ${slo.indicator.type} SLO type`);
       throw new Error(`Unsupported SLI type: ${slo.indicator.type}`);
     }
 
-    const transformParams = generator.getTransformParams(slo);
+    const transformParams = rollupGenerator.getTransformParams(slo);
     try {
       await retryTransientEsErrors(() => this.esClient.transform.putTransform(transformParams), {
         logger: this.logger,
       });
     } catch (err) {
-      this.logger.error(`Cannot create transform for ${slo.indicator.type} SLI type: ${err}`);
+      this.logger.error(`Cannot create transform: ${err}`);
       throw err;
     }
